First complete draft of proc unit testing code.
Most mocked functions utilize mock_fn(), which checks for expected input
and either returns the appropriate output given the mocking or lets the
original function be called.  This is necessary in cases where outside
code calls these functions and the mocked version is not able to supply
them with the needed output.  Private methods in the stem/util/proc.py
file are not tested, nor is is_available().  This decision was based on
the conventions set in system.py testing code.